style: Move DocumentStyleSheet to style.
